moved filed, renamed variables
