+++ src/app/containers/post/CommentsContainer.jsx
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
 
-import CommentsHeader from 'src/app/containers/post/CommentsHeader';
-import CreateComment from 'src/app/containers/post/CreateComment';
-import CommentsList from 'src/app/containers/post/CommentsList';
+import CommentsHeader from 'src/app/components/post/CommentsHeader';
+import CreateComment from 'src/app/components/post/CreateComment';
+import CommentsList from 'src/app/components/post/CommentsList';
 import commentsContainerSelector from 'src/app/redux/selectors/post/commentsContainer';
 import { fetchCommentsIfNeeded } from 'src/app/redux/actions/comments';
