+++ src/app/containers/post/PostContainer.jsx
@@ -4,14 +4,14 @@ import styled from 'styled-components';
 
 import Container from 'src/app/components/common/Container/Container';
 import SidePanel from 'src/app/containers/post/SidePanel';
-import PostContent from 'src/app/components/post/PostContent';
+import PostContent from 'src/app/containers/post/PostContent';
 import ActivePanel from 'src/app/containers/post/ActivePanel';
 import AboutPanel from 'src/app/containers/post/AboutPanel';
 import { USER_FOLLOW_DATA_LOAD } from 'src/app/redux/constants/followers';
 import { FAVORITES_LOAD } from 'src/app/redux/constants/favorites';
 import LoadingIndicator from 'app/components/elements/LoadingIndicator';
 import CommentsContainer from 'src/app/containers/post/CommentsContainer';
-import RegistrationPanel from 'src/app/containers/post/RegistrationPanel';
+import RegistrationPanel from 'src/app/components/post/RegistrationPanel';
 import { postContainerSelector } from 'src/app/redux/selectors/post/postContainer';
 
 const Wrapper = styled.div`
